Try 1: GitHub Action for `Resume Generate`
This should use XeLaTeX to generate a PDF of my resume to the repo by sourcing the code present. This should take place on push.

Concerns:
Breadth of scope: will it see the entire folder? There are references further into the directory used in the main folder.
Bloat: Will it delete previous versions?